Moved examples to top level
demo is also now the bevy example
