+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["crates/*"]
+members = ["crates/*", "examples/*"]
 
 [workspace.package]
 version = "0.1.0"
@@ -12,10 +12,9 @@ repository = "https://github.com/urholaukkarinen/transform-gizmo"
 authors = ["Urho Laukkarinen <urho.laukkarinen@gmail.com>"]
 
 [workspace.dependencies]
-transform-gizmo = { path = "crates/transform-gizmo", default-features = false }
-transform-gizmo-egui = { path = "crates/transform-gizmo-egui", default-features = false }
-transform-gizmo-bevy = { path = "crates/transform-gizmo-bevy", default-features = false }
-transform-gizmo-demo = { path = "crates/transform-gizmo-demo", default-features = false }
+transform-gizmo = { path = "crates/transform-gizmo" }
+transform-gizmo-egui = { path = "crates/transform-gizmo-egui" }
+transform-gizmo-bevy = { path = "crates/transform-gizmo-bevy" }
 
 egui = "0.27.2"
 eframe = "0.27.2"
