--- crates/transform-gizmo-demo/.gitignore
@@ -1,1 +1,0 @@
-target/
