--- crates/transform-gizmo-demo/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "transform-gizmo-demo"
-description = "demo app for transform-gizmo"
-
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-transform-gizmo-bevy.workspace = true
-transform-gizmo-egui.workspace = true
-
-bevy.workspace = true
-bevy_infinite_grid = { git = "https://github.com/XYCaptain/bevy_infinite_grid.git", rev = "1e5259a52a7dfaca65402c2c57c1d81c2bc443e9" }
-bevy_mod_picking = "0.18.0"
-bevy_mod_outline = "0.7.0"
-
-[dependencies.bevy_egui]
-git = "https://github.com/mvlabat/bevy_egui.git"
-rev = "cfd819ee9bdfb58101ac3efb6bcd4b6336860773"
-features = ["open_url", "default_fonts", "render"]
-default-features = false
-
-[lints]
-workspace = true
