--- crates/transform-gizmo-demo/Trunk.toml
@@ -1,3 +1,0 @@
-[build]
-filehash = false
-dist = "../../docs"
