--- crates/transform-gizmo-demo/index.html
@@ -1,43 +1,0 @@
-<html>
-
-<head>
-    <title>transform-gizmo demo</title>
-    <style>
-        html {
-            /* Remove touch delay: */
-            touch-action: manipulation;
-        }
-
-        html,
-        body,
-        canvas {
-            margin: 0 !important;
-            padding: 0 !important;
-            height: 100% !important;
-            width: 100% !important;
-            overflow: hidden;
-            background: black;
-            z-index: 0;
-        }
-
-        /* Position canvas in center-top: */
-        canvas {
-            outline: none;
-            display: block;
-            position: absolute;
-            top: 0%;
-            left: 50%;
-            transform: translate(-50%, 0%);
-        }
-    </style>
-
-    <link data-trunk rel="rust" href="Cargo.toml" data-wasm-opt="z" data-bin="transform-gizmo-demo" />
-
-    <!-- this is the base url relative to which other urls will be constructed. trunk will insert this from the public-url option -->
-    <base data-trunk-public-url />
-</head>
-
-<body oncontextmenu="return false;">
-</body>
-
-</html>
