+++ docs/index.html
@@ -1,0 +1,1 @@
+<title>transform-gizmo demo</title><style>html{touch-action:manipulation}body,canvas,html{margin:0!important;padding:0!important;height:100%!important;width:100%!important;overflow:hidden;background:black;z-index:0}canvas{outline:none;display:block;position:absolute;top:0%;left:50%;transform:translate(-50%,0%)}</style><script type=module>import a,*as b from"/bevy-example.js";a(`/bevy-example_bg.wasm`);window.wasmBindings=b</script><base href=/><link as=fetch crossorigin href=/bevy-example_bg.wasm integrity=sha384-6Z-SFN6kAgL5W3NCtJ7RqA9nD8_l2w1hxCKpL31AvyOqE4EMrTdglfKmKt-KNNO2 rel=preload type=application/wasm><link crossorigin href=/bevy-example.js integrity=sha384-8IAkFedG4DooakeSIaa7WF3mzHDL32PU-TojrUtEWAND2K62-AwdQhSAZBbCAImk rel=modulepreload></head><body oncontextmenu="return false;"></body></html>
