+++ examples/bevy/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "bevy-example"
+version.workspace = true
+rust-version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+authors.workspace = true
+publish = false
+
+[dependencies]
+transform-gizmo-bevy.workspace = true
+
+bevy.workspace = true
+bevy_infinite_grid = { git = "https://github.com/XYCaptain/bevy_infinite_grid.git", rev = "1e5259a52a7dfaca65402c2c57c1d81c2bc443e9" }
+bevy_mod_picking = "0.18.0"
+bevy_mod_outline = "0.7.0"
+
+[dependencies.bevy_egui]
+git = "https://github.com/mvlabat/bevy_egui.git"
+rev = "cfd819ee9bdfb58101ac3efb6bcd4b6336860773"
+features = ["open_url", "default_fonts", "render"]
+default-features = false
+
+[lints]
+workspace = true
