+++ examples/bevy/Trunk.toml
@@ -1,0 +1,3 @@
+[build]
+filehash = false
+dist = "../../docs"
