+++ examples/egui/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "egui-example"
+version.workspace = true
+rust-version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+authors.workspace = true
+publish = false
+
+[dependencies]
+transform-gizmo-egui.workspace = true
+eframe.workspace = true
+
+[lints]
+workspace = true
